William Henry B05 Monarch
The Monarch "Departure" has fossil coral scales and hand carved sterling silver skulls on the bolsters! The Hornets Nest damascus blade, hand forged by Mike Norris, has a beautiful pattern to it! The thumb stud and the button lock are set with sapphire! This one is number 46 of 100 made as pictured!
The William Henry Monarch is a small gentleman's knife perfect for carrying in slacks or a suit. The knife opens one handed with the thumb stud and unlocks with a button lock. It comes with a leather carrying case that has a clip that can be clipped so that the knife sits on the inside of a pocket. This keeps the knife protected, and the knife is much more comfortable in the hand without a pocket clip on the side of the knife. It also has a braided leather lanyard to make it easy to pull the knife out of the carrying case.